feat: remove not needed DeadlineInput component
