--- src/cow-react/modules/advancedOrders/containers/DeadlineInput/styled.tsx
@@ -1,6 +1,0 @@
-import styled from 'styled-components/macro'
-
-export const LabelWrapper = styled.div`
-  display: flex;
-  align-items: center;
-`
